config: Change the worker start delay to 10msec
It seems the two session limitation from the same IP address
is for real.  I think I can go down the delay to 1msec but
just to be on the conservative side for now.  As the detection
is source IP address base, I need to move on to the distributed
system to overcome the API limitation, which comes down to the
distributed system, which runs the worker instance on a different
machines and use the TCP socket to report the result, instaed
of using the current std::sync::mpsc channel.  For this reason,
I'll keep the current mpsc based interaction between Tarcker
and Worker to make it easier to progress.

Next, I'll use indicatif::MultiProgress to show the realtime
track update.